Added resources py file which fixes uninitalized module bug when creating a resume.
