+++ website/__init__.py
@@ -1,41 +1,22 @@
 from flask import (
     Flask,
     render_template,
-    redirect, 
-    request,
+    redirect,
     escape
 )
-import json
-import requests
-
-GITHUB_REPO_LIST_URL = 'https://api.github.com/users/santosderek/repos?page={number}'
-GITHUB_USER_URL = 'https://api.github.com/users/santosderek'
-
-
-class GithubRequestError(Exception):
-    pass
-
-
-def get_github_user():
-    """Requests from the GITHUB API my user data and returns as JSON(Dict)"""
-    returned_value = requests.get(GITHUB_USER_URL)
-
-    if returned_value.status_code != 200:
-        raise GithubRequestError("Error Code != 200")
-
-    return returned_value.json()
-
-
-def get_resource_json(filename: str):
-    "Return the contents of a resource JSON file."
-    with open(f'website/resources/{filename}', 'r') as current_file:
-        data = json.loads(current_file.read())
-        return data
+from .resources import get_resource_json, GithubRequestError, get_github_user
+from .resume import generate_document, DOWNLOAD_LOCATION
+from os.path import exists
 
 
 def create_app():
     app = Flask(__name__)
     app.config.from_pyfile('settings.py')
+
+    app.logger.info("Creating the resume.")
+    generate_document()
+    assert exists(DOWNLOAD_LOCATION)
+    app.logger.info("Resume Created.")
 
     @app.route('/', methods=["GET"])
     def home():
@@ -77,6 +58,12 @@ def create_app():
                                careers=careers,
                                educations=educations,
                                repos=repos)
+
+    @app.route('/resume', methods=["GET"])
+    def resume():
+        """This route returns a download of my resume."""
+
+        return redirect('https://www.linkedin.com/in/santosderek/')
 
     @app.route('/project/<string:project>', methods=["GET"])
     def project(project: str):
